// src/services/models/memeControllerQueryPaginationFilter.ts
/**
 * Generated by orval v7.1.0 🍺
 * Do not edit manually.
 * A8 Pump Fun API
 * OpenAPI spec version: 0.0.12
 */

export type MemeControllerQueryPaginationFilter =
  (typeof MemeControllerQueryPaginationFilter)[keyof typeof MemeControllerQueryPaginationFilter];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const MemeControllerQueryPaginationFilter = {
  all: "all",
  end: "end",
  active: "active",
} as const;