// src/services/models/memeControllerQueryPaginationSortBy.ts
/**
 * Generated by orval v7.1.0 🍺
 * Do not edit manually.
 * A8 Pump Fun API
 * OpenAPI spec version: 0.0.12
 */

export type MemeControllerQueryPaginationSortBy =
  (typeof MemeControllerQueryPaginationSortBy)[keyof typeof MemeControllerQueryPaginationSortBy];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const MemeControllerQueryPaginationSortBy = {
  createdTime: "createdTime",
  volume24hs: "volume24hs",
  marketCap: "marketCap",
} as const;