// src/services/models/memeControllerQueryPaginationSortOrder.ts
/**
 * Generated by orval v7.1.0 🍺
 * Do not edit manually.
 * A8 Pump Fun API
 * OpenAPI spec version: 0.0.12
 */

export type MemeControllerQueryPaginationSortOrder =
  (typeof MemeControllerQueryPaginationSortOrder)[keyof typeof MemeControllerQueryPaginationSortOrder];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const MemeControllerQueryPaginationSortOrder = {
  DESC: "DESC",
  ASC: "ASC",
} as const;